chore: refactor with BaseCell
