+++ JWords/iOSView/items/StudyCell.swift
@@ -31,64 +31,8 @@ struct StudyWord: ReducerProtocol {
             self.word = word
             self.isLocked = false
             self.frontType = frontType
-            self.studyState = word.studyState
         }
-        
-        var frontText: String {
-            switch frontType {
-            case .meaning:
-                return word.meaningText
-            case .kanji:
-                return word.kanjiText
-            }
-        }
-        
-        var frontImageURLs: [URL] {
-            switch frontType {
-            case .meaning:
-                return [word.meaningImageURL]
-                    .filter { !$0.isEmpty }
-                    .compactMap { URL(string: $0) }
-            case .kanji:
-                return [word.kanjiImageURL]
-                    .filter { !$0.isEmpty }
-                    .compactMap { URL(string: $0) }
-            }
-        }
-        
-        // frontText를 제외한 두 가지 text에서 빈 text를 제외하고 띄어쓰기
-        var backText: String {
-            switch frontType {
-            case .meaning:
-                return [word.ganaText, word.kanjiText]
-                    .filter { !$0.isEmpty }
-                    .joined(separator: "\n")
-            case .kanji:
-                return [word.ganaText, word.meaningText]
-                    .filter { !$0.isEmpty }
-                    .joined(separator: "\n")
-            }
-        }
-        
-        var backImageURLs: [URL] {
-            switch frontType {
-            case .meaning:
-                return [word.kanjiImageURL, word.ganaImageURL]
-                    .filter { !$0.isEmpty }
-                    .compactMap { URL(string: $0) }
-            case .kanji:
-                return [word.ganaImageURL, word.meaningImageURL]
-                    .filter { !$0.isEmpty }
-                    .compactMap { URL(string: $0) }
-            }
-            
-        }
-        
-        static func == (lhs: StudyWord.State, rhs: StudyWord.State) -> Bool {
-            lhs.id == rhs.id
-            && lhs.isFront == rhs.isFront
-            && lhs.studyState == rhs.studyState
-        }
+
     }
     
     enum SwipeDirection: Equatable {
@@ -140,7 +84,6 @@ struct StudyCell: View {
     let store: StoreOf<StudyWord>
     
     @GestureState private var dragAmount = CGSize.zero
-    @State private var deviceWidth: CGFloat = Constants.Size.deviceWidth
     
     // gestures
     let dragGesture = DragGesture(minimumDistance: 30, coordinateSpace: .global)
@@ -150,98 +93,21 @@ struct StudyCell: View {
     // MARK: Body
     var body: some View {
         WithViewStore(store, observe: { $0 }) { vs in
-            ZStack {
-                sizeDecisionView(frontText: vs.frontText,
-                                 frontImageURLs: vs.frontImageURLs,
-                                 backText: vs.backText,
-                                 backImageURLs: vs.backImageURLs)
-                swipeGuide
-                ZStack {
-                    cellColor(vs.studyState)
-                    cellFace(vs.isFront ? vs.frontText : vs.backText,
-                             vs.isFront ? vs.frontImageURLs : vs.backImageURLs)
-                }
-                .offset(dragAmount)
-            }
-            .frame(width: deviceWidth * 0.9)
-            .frame(minHeight: vs.word.hasImage ? 200 : 100)
+            BaseCell(word: vs.word,
+                     frontType: vs.frontType,
+                     isFront: vs.isFront,
+                     dragAmount: dragAmount)
             .gesture(dragGesture
                 .updating($dragAmount) { dragUpdating(vs.isLocked, $0, &$1, &$2) }
                 .onEnded { vs.send(.cellDrag(direction: $0.translation.width > 0 ? .left : .right)) }
             )
             .gesture(doubleTapGesture.onEnded { vs.send(.cellDoubleTapped) })
             .gesture(tapGesture.onEnded { vs.send(.cellTapped) })
-            #if os(iOS)
-            .onAppear { deviceOrientationChanged() }
-            .onReceive(NotificationCenter.default.publisher(for: UIDevice.orientationDidChangeNotification)) { _ in deviceOrientationChanged() }
-            #endif
         }
     }
     
 }
 
-// MARK: SubViews
-
-extension StudyCell {
-    
-    private func sizeDecisionView(frontText: String,
-                                  frontImageURLs: [URL],
-                                  backText: String,
-                                  backImageURLs: [URL]) -> some View {
-        ZStack {
-            ZStack {
-                cellFace(frontText, frontImageURLs)
-                Color.white
-            }
-            ZStack {
-                cellFace(backText, backImageURLs)
-                Color.white
-            }
-        }
-    }
-    
-    private var swipeGuide: some View {
-        HStack {
-            Image(systemName: "circle")
-                .resizable()
-                .frame(width: 100, height: 100)
-                .foregroundColor(.blue)
-            Spacer()
-            Image(systemName: "x.circle")
-                .resizable()
-                .frame(width: 100, height: 100)
-                .foregroundColor(.red)
-        }
-        .background { Color.white }
-    }
-    
-    private func cellColor(_ studyState: StudyState) -> some View {
-        Group {
-            switch studyState {
-            case .undefined:
-                Color.white
-            case .success:
-                Color(red: 207/256, green: 240/256, blue: 204/256)
-            case .fail:
-                Color(red: 253/256, green: 253/256, blue: 150/256)
-            }
-        }
-    }
-    
-    private func cellFace(_ text: String, _ imageURLs: [URL]) -> some View {
-        VStack {
-            Text(text)
-                .font(.system(size: fontSize(of: text)))
-            VStack {
-                ForEach(imageURLs, id: \.self) { url in
-                    KFImage(url)
-                        .resizable()
-                        .scaledToFit()
-                }
-            }
-        }
-    }
-}
 
 // MARK: View Methods
 
@@ -252,21 +118,6 @@ extension StudyCell {
         state.width = value.translation.width
     }
     
-    private func deviceOrientationChanged() {
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-            self.deviceWidth = Constants.Size.deviceWidth
-        }
-    }
-    
-    private func fontSize(of text: String) -> CGFloat {
-        if text.count <= 10 {
-            return 45
-        } else if text.count <= 30 {
-            return 35
-        } else {
-            return 30
-        }
-    }
 }
 
 // MARK: ViewModel
